chore(performance): Rename folder
