+++ src/sentry/testutils/performance_issues/event_generators.py
