+++ src/sentry/testutils/performance_issues/events/fast-n-plus-one-in-django-new-view.json
