+++ src/sentry/testutils/performance_issues/events/n-plus-one-in-django-index-view-activerecord.json
