+++ src/sentry/testutils/performance_issues/events/n-plus-one-in-django-new-view.json
