+++ src/sentry/testutils/performance_issues/events/n-plus-one-in-rails-index-view.json
