+++ src/sentry/testutils/performance_issues/events/no-issue-in-django-detail-view.json
