+++ src/sentry/testutils/performance_issues/events/query-waterfall-in-django-random-view.json
