+++ src/sentry/testutils/performance_issues/events/solved-n-plus-one-in-django-index-view.json
